tagsheet/textwidget: method for applying inlinetags that sets less options in the text widget. This causes problems when the generated tags are not deleted upon switching to a different tagsheet.
Consider tagsheet A:
	inlinetag myinlinetag {size += 4}
and tagsheet B:
	inlinetag myinlinetag {color = blue}
When a tag is generated in the same text widget, first from tagsheet A,
then from tagsheet B:
	inlinetag_configure $w $tagsheetA : tag = myinlinetag PARENT r
	inlinetag_configure $w $tagsheetB : tag = myinlinetag PARENT r
the "-font" option is still set to a certain value corresponding to
tagsheet A. Conversely, when switching from tagsheet B to A, the tag
will still have a "-color" option set to blue.

So, you must make sure that you delete all tag information before
reconfiguring tags from a new tagsheet. As this can lead to hard-to-find
bugs, the "more efficient" option-setting code only lives here in its
own branch.